+++ crates/rspack_plugin_javascript/src/plugin.rs
@@ -67,8 +67,8 @@ impl ParserAndGenerator for JavaScriptParserAndGenerator {
     SOURCE_TYPES
   }
 
-  fn size(&self, module: &NormalModule, _source_type: &SourceType) -> f32 {
-    module.original_source().map_or(0, |source| source.size()) as f32
+  fn size(&self, module: &NormalModule, _source_type: &SourceType) -> f64 {
+    module.original_source().map_or(0, |source| source.size()) as f64
   }
 
   #[instrument(name = "js:parse")]
